# apps/api_v2/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api_v2",
    "description": "Public markdown content, categories and search text, and employee-authenticated editor operations. Editor reads, writes, and the version token require a GitHub-authenticated govtech-bb organization member session.",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "previewToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-preview-token",
        "description": "The site's PREVIEW_SECRET, sent server to server. Optional on site reads; with it, `preview` content is served too."
      },
      "editorSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "HTTP-only employee session cookie; Secure-prefixed in production."
      }
    },
...
  },
  "paths": {
...
    "/categories": {
      "get": {
        "summary": "The categories the site lists",
        "tags": [
          "categories"
        ],
        "description": "In order, each with its subcategories. A category is listed when it or one of its subcategories has a visible page at its root.",
        "security": [
          {},
          {
            "previewToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string",
                            "description": "`/<category>`, or `/<category>/<subcategory>`."
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          },
                          "subcategories": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "slug": {
                                  "type": "string"
                                },
                                "url": {
                                  "type": "string",
                                  "description": "`/<category>`, or `/<category>/<subcategory>`."
                                },
                                "title": {
                                  "type": "string"
                                },
                                "description": {
                                  "nullable": true,
                                  "type": "string"
                                }
                              },
                              "required": [
                                "slug",
                                "url",
                                "title",
                                "description"
                              ],
                              "additionalProperties": false
                            }
                          }
                        },
                        "required": [
                          "slug",
                          "url",
                          "title",
                          "description",
                          "subcategories"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "categories"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/categories/{slug}": {
      "get": {
        "summary": "A category and what it lists",
        "tags": [
          "categories"
        ],
        "description": "Its subcategories and the visible pages at its root. 404 when it lists nothing.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {},
          {
            "previewToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "category": {
                      "type": "object",
                      "properties": {
                        "slug": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string",
                          "description": "`/<category>`, or `/<category>/<subcategory>`."
                        },
                        "title": {
                          "type": "string"
                        },
                        "description": {
                          "nullable": true,
                          "type": "string"
                        }
                      },
                      "required": [
                        "slug",
                        "url",
                        "title",
                        "description"
                      ],
                      "additionalProperties": false
                    },
                    "parent": {
                      "nullable": true,
                      "description": "The category this is a subcategory of, or null.",
                      "type": "object",
                      "properties": {
                        "slug": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string",
                          "description": "`/<category>`, or `/<category>/<subcategory>`."
                        },
                        "title": {
                          "type": "string"
                        },
                        "description": {
                          "nullable": true,
                          "type": "string"
                        }
                      },
                      "required": [
                        "slug",
                        "url",
                        "title",
                        "description"
                      ],
                      "additionalProperties": false
                    },
                    "subcategories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string",
                            "description": "`/<category>`, or `/<category>/<subcategory>`."
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          }
                        },
                        "required": [
                          "slug",
                          "url",
                          "title",
                          "description"
                        ],
                        "additionalProperties": false
                      },
                      "description": "Those with something to list, in order."
                    },
                    "pages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          },
                          "digital": {
                            "type": "boolean",
                            "description": "It has a form, or its frontmatter says it is digital."
                          }
                        },
                        "required": [
                          "url",
                          "title",
                          "description",
                          "digital"
                        ],
                        "additionalProperties": false
                      },
                      "description": "The pages at the category's root, A to Z."
                    }
                  },
                  "required": [
                    "category",
                    "parent",
                    "subcategories",
                    "pages"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/categories/{category}/{slug}": {
      "get": {
        "summary": "A subcategory and what it lists",
        "tags": [
          "categories"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "category",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {},
          {
            "previewToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "category": {
                      "type": "object",
                      "properties": {
                        "slug": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string",
                          "description": "`/<category>`, or `/<category>/<subcategory>`."
                        },
                        "title": {
                          "type": "string"
                        },
                        "description": {
                          "nullable": true,
                          "type": "string"
                        }
                      },
                      "required": [
                        "slug",
                        "url",
                        "title",
                        "description"
                      ],
                      "additionalProperties": false
                    },
                    "parent": {
                      "nullable": true,
                      "description": "The category this is a subcategory of, or null.",
                      "type": "object",
                      "properties": {
                        "slug": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string",
                          "description": "`/<category>`, or `/<category>/<subcategory>`."
                        },
                        "title": {
                          "type": "string"
                        },
                        "description": {
                          "nullable": true,
                          "type": "string"
                        }
                      },
                      "required": [
                        "slug",
                        "url",
                        "title",
                        "description"
                      ],
                      "additionalProperties": false
                    },
                    "subcategories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string",
                            "description": "`/<category>`, or `/<category>/<subcategory>`."
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          }
                        },
                        "required": [
                          "slug",
                          "url",
                          "title",
                          "description"
                        ],
                        "additionalProperties": false
                      },
                      "description": "Those with something to list, in order."
                    },
                    "pages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          },
                          "digital": {
                            "type": "boolean",
                            "description": "It has a form, or its frontmatter says it is digital."
                          }
                        },
                        "required": [
                          "url",
                          "title",
                          "description",
                          "digital"
                        ],
                        "additionalProperties": false
                      },
                      "description": "The pages at the category's root, A to Z."
                    }
                  },
                  "required": [
                    "category",
                    "parent",
                    "subcategories",
                    "pages"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/catalog": {
      "get": {
        "summary": "Every visible page",
        "tags": [
          "pages"
        ],
        "description": "Sub-pages included, `start` steps not, A to Z: for the sitemap and service lists.",
        "security": [
          {},
          {
            "previewToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          },
                          "digital": {
                            "type": "boolean",
                            "description": "It has a form, or its frontmatter says it is digital."
                          },
                          "stage": {
                            "nullable": true,
                            "type": "string"
                          }
                        },
                        "required": [
                          "url",
                          "title",
                          "description",
                          "digital",
                          "stage"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "pages"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}